# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Solar SCADA IoT Monitoring & Control System API
 * OpenAPI spec version: 0.1.0
 */
export interface GeminiConversation {
  id: number;
  title: string;
  createdAt: string;
}

export interface GeminiMessage {
  id: number;
  conversationId: number;
  role: string;
  content: string;
  createdAt: string;
}

export interface GeminiNewConversation {
  title: string;
}

export interface GeminiNewMessage {
  content: string;
}

export interface GeminiConversationWithMessages {
  id: number;
  title: string;
  createdAt: string;
  messages: GeminiMessage[];
}

export interface GeminiImagePrompt {
  prompt: string;
}

export interface GeminiImageResult {
  b64_json: string;
  mimeType: string;
}

export interface GeminiError {
  error: string;
}

export interface HealthStatus {
  status: string;
}

export type SensorReadingSystemStatus =
  (typeof SensorReadingSystemStatus)[keyof typeof SensorReadingSystemStatus];

export const SensorReadingSystemStatus = {
  normal: "normal",
  warning: "warning",
  fault: "fault",
  offline: "offline",
} as const;

export interface SensorReading {
  id: number;
  /** Solar panel voltage in Volts */
  voltage: number;
  /** Current in Amperes */
  current: number;
  /** Power output in Watts */
  power: number;
  /** Battery charge percentage (0-100) */
  batteryLevel: number;
  /** Battery voltage in Volts */
  batteryVoltage: number;
  /** Panel temperature in Celsius */
  temperature: number;
  /** Solar irradiance in W/m² */
  irradiance: number;
  /** Load/consumption power in Watts */
  loadPower: number;
  systemStatus: SensorReadingSystemStatus;
  timestamp: string;
}

export type SensorReadingInputSystemStatus =
  (typeof SensorReadingInputSystemStatus)[keyof typeof SensorReadingInputSystemStatus];

export const SensorReadingInputSystemStatus = {
  normal: "normal",
  warning: "warning",
  fault: "fault",
  offline: "offline",
} as const;

export interface SensorReadingInput {
  voltage: number;
  current: number;
  power: number;
  batteryLevel: number;
  batteryVoltage: number;
  temperature: number;
  irradiance: number;
  loadPower: number;
  systemStatus?: SensorReadingInputSystemStatus;
}

export interface HistoricalPoint {
  timestamp: string;
  value: number;
  metric: string;
}

export type AlertType = (typeof AlertType)[keyof typeof AlertType];

export const AlertType = {
  overvoltage: "overvoltage",
  undervoltage: "undervoltage",
  overtemperature: "overtemperature",
  low_battery: "low_battery",
  fault: "fault",
  info: "info",
} as const;

export type AlertSeverity = (typeof AlertSeverity)[keyof typeof AlertSeverity];

export const AlertSeverity = {
  critical: "critical",
  warning: "warning",
  info: "info",
} as const;

export type AlertStatus = (typeof AlertStatus)[keyof typeof AlertStatus];

export const AlertStatus = {
  active: "active",
  resolved: "resolved",
} as const;

export interface Alert {
  id: number;
  type: AlertType;
  severity: AlertSeverity;
  message: string;
  status: AlertStatus;
  timestamp: string;
  resolvedAt?: string | null;
}

export type DeviceType = (typeof DeviceType)[keyof typeof DeviceType];

export const DeviceType = {
  inverter: "inverter",
  charge_controller: "charge_controller",
  load: "load",
  sensor: "sensor",
  pump: "pump",
} as const;

export type DeviceStatus = (typeof DeviceStatus)[keyof typeof DeviceStatus];

export const DeviceStatus = {
  on: "on",
  off: "off",
  fault: "fault",
} as const;

export interface Device {
  id: number;
  name: string;
  type: DeviceType;
  status: DeviceStatus;
  isEnabled: boolean;
  /** Power rating in Watts */
  powerRating: number;
  location: string;
}

export interface DeviceToggleInput {
  isEnabled: boolean;
}

export interface LogEntry {
  id: number;
  timestamp: string;
  voltage: number;
  current: number;
  power: number;
  batteryLevel: number;
  temperature: number;
  irradiance: number;
  loadPower: number;
  systemStatus: string;
}

export interface LogsResponse {
  data: LogEntry[];
  total: number;
  page: number;
  totalPages: number;
}

export type DashboardSummarySystemStatus =
  (typeof DashboardSummarySystemStatus)[keyof typeof DashboardSummarySystemStatus];

export const DashboardSummarySystemStatus = {
  normal: "normal",
  warning: "warning",
  fault: "fault",
  offline: "offline",
} as const;

export interface DashboardSummary {
  currentPower: number;
  /** kWh generated today */
  energyToday: number;
  /** Total kWh generated */
  energyTotal: number;
  batteryLevel: number;
  systemStatus: DashboardSummarySystemStatus;
  activeAlerts: number;
  devicesOnline: number;
  totalDevices: number;
  /** Cost savings in USD today */
  savingsToday: number;
  /** CO2 saved in kg */
  co2Saved: number;
}

export interface EnergyHourPoint {
  hour: string;
  production: number;
  consumption: number;
}

export interface AlertsSummary {
  total: number;
  critical: number;
  warning: number;
  info: number;
  active: number;
  resolved: number;
}

export type GetReadingsHistoryParams = {
  hours?: number;
  metric?: GetReadingsHistoryMetric;
};

export type GetReadingsHistoryMetric =
  (typeof GetReadingsHistoryMetric)[keyof typeof GetReadingsHistoryMetric];

export const GetReadingsHistoryMetric = {
  voltage: "voltage",
  current: "current",
  power: "power",
  battery: "battery",
  temperature: "temperature",
  irradiance: "irradiance",
} as const;

export type ListAlertsParams = {
  status?: ListAlertsStatus;
};

export type ListAlertsStatus =
  (typeof ListAlertsStatus)[keyof typeof ListAlertsStatus];

export const ListAlertsStatus = {
  active: "active",
  resolved: "resolved",
  all: "all",
} as const;

export type ListLogsParams = {
  limit?: number;
  page?: number;
};