# lib/api-zod/src/generated/types/geminiNewMessage.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Solar SCADA IoT Monitoring & Control System API
 * OpenAPI spec version: 0.1.0
 */

export interface GeminiNewMessage {
  content: string;
}